UI Streamlit/requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
opencv-python>=4.8.0
joblib>=1.3.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pillow>=10.0.0
